meta-tools/gamification/requirements.txt
# Core dependencies for gamification tools
requests>=2.31.0          # For API calls (Habitica, AI providers)
matplotlib>=3.7.0         # For charts and visualizations
plotly>=5.14.0           # For interactive charts
pandas>=2.0.0            # For data analysis
numpy>=1.24.0            # For numerical operations
networkx>=3.1            # For network graphs
pyvis>=0.3.2             # For network visualizations
jinja2>=3.1.0            # For HTML report generation
python-dotenv>=1.0.0     # For loading environment variables

# Optional dependencies for enhanced features
# flask>=3.0.0           # For web-based stats viewer (if needed)
# sqlite3                # Built-in Python module for productivity tracking